requirements.txt
# Core dependencies
streamlit>=1.33.0
python-dotenv>=1.0.1

# LangChain dependencies
langchain>=0.1.16
langchain-community>=0.0.34
langchain-openai>=0.1.4

# OpenAI
openai>=1.24.0

# PDF processing
pypdf>=4.2.0

# Vector database
faiss-cpu>=1.8.0

# Image processing
Pillow>=9.0.0

# Development (optional)
ipykernel>=6.29.4